Добавлен шаблон корзины, кнопка для авторизованных пользователей
