+++ app/shop/views.py
@@ -1,4 +1,5 @@
 from flask import render_template, Blueprint
+from app.shop.models import Product
 
 blueprint = Blueprint('shop', __name__, url_prefix='/shop', template_folder='templates/shop')
 
@@ -8,7 +9,16 @@ blueprint = Blueprint('shop', __name__, url_prefix='/shop', template_folder='tem
 @blueprint.route("/")
 def shop():
     page_title = 'Магазин'
-    content = 'Страница магазина'
-    products = product_list
+    products_list = Product.query.order_by(Product.id).all()
     return render_template('shop/shop.html',
-                           page_title=page_title, content=content, products=products)
+                           page_title=page_title, products=products_list, contetn="Товары")
+
+
+
+# @blueprint.route('/<product_id>')
+# def product():
+#     products_list = Product.query.order_by(Product.id).all
+#     title = f"Страница продукта {products_list['id']}"
+#     content = 'Страничка продукта'
+#     return render_template('shop/product.html',
+#                            page_title=title, content=content)
