+++ app/templates/carts/cart.html
@@ -1,0 +1,66 @@
+{% extends 'base.html' %}
+
+{% block title %}
+{{ page_title }}
+{% endblock %}
+
+
+{% block content %}
+<div class="row justify-content-center">
+    <div class=" bg-white p-4 mb-4 mx-2 rounded custom-shadow">
+        <h3 class="text-center mb-4">Корзина</h3>
+        <div class="container" id="cart-items-container">
+            <!-- Разметка корзины -->
+            <div class="card mb-3 text-bg-light shadow-lg">
+                <div class="card-header">
+                    <h5 class="card-title">Товар</h5>
+                </div>
+                <ul class="list-group list-group-flush">
+                    <li class="list-group-item">
+                        <div class="row text-center">
+                            <div class="col p-0">
+                                <div class="input-group">
+                                    <span class="input-group-btn">
+                                        <button type="button"
+                                                class="btn btn-primary btn-sm decrement"
+                                                data-cart-id=""
+                                                data-cart-change-url="">-</button>
+                                    </span>
+                                    <input type="text" class="form-control number" value="1" readonly="">
+                                    <span class="input-group-btn">
+                                        <button type="button"
+                                                class="btn btn-primary btn-sm increment"
+                                                data-cart-id=""
+                                                data-cart-change-url="">+</button>
+                                    </span>
+                                </div>
+                            </div>
+                            <div class="col p-0">
+                                <p>x 100.00 = </p>
+                            </div>
+                            <div class="col p-0"><strong>100.00 $</strong></div>
+                            <div class="col p-0">
+                                <a href="#" class="remove-from-cart" data-cart-id="">
+                                    <img class="mx-1" src="" alt="Catalog Icon" width="16"
+                                         height="16">
+                                </a>
+                            </div>
+                        </div>
+                    </li>
+                </ul>
+            </div>
+            <div class="card mb-3 shadow-lg">
+                <div class="card-footer">
+                    <p class="float-left">Итого <strong>1</strong> товар(а) на сумму</p>
+                    <h4 class="float-left"><strong>100.00 $</strong></h4>
+                </div>
+            </div>
+            <!-- Закончилась разметка корзины -->
+        </div>
+        <a class="btn btn-primary" href="">
+            Оформить заказ
+        </a>
+    </div>
+</div>
+{{ content }}
+{% endblock %}
