+++ app/templates/shop/shop.html
@@ -16,6 +16,7 @@
                 <a href="#" class="btn btn-primary my-2">Все товары</a>
                 <a href="#" class="btn btn-primary my-2">Для женщин</a>
                 <a href="#" class="btn btn-primary my-2">Для мужчин</a>
+                <a href="#" class="btn btn-primary my-2">Для детей</a>
             </p>
         </div>
     </div>
@@ -23,44 +24,29 @@
 
 <div class="row row-cols-1 row-cols-sm-2 row-cols-md-3 g-3">
     {% for product in products %}
-    <div class="col">
-        <div class="card border-primary rounded custom-shadow">
-            <img src="../deps/images/goods/set of tea table and three chairs.jpg" class="card-img-top" alt="...">
-            <div class="card-body">
-                <a href="../goods/product.html">
-                    <p class="card-title">{{ product.name }}</p>
-                </a>
-                <p class="card-text text-truncate">{{ product.description }}</p>
-                <p class="product_id">{{ product.id }}</p>
-                <div class="d-flex justify-content-between">
-                    <p><s>100</s> $</p>
-                    <p><strong>90.00 $</strong></p>
-                    <span class="badge bg-warning text-dark">Скидка 10.00 %</span>
-                    <!-- <p><strong>100.00 $</strong></p> -->
-                    <a href="#" class="btn add-to-cart">
-                        <img class="mx-1" src="../deps/icons/cart-plus.svg" alt="Catalog Icon" width="32" height="32">
+        <div class="col">
+            <div class="card border-primary rounded custom-shadow">
+                <img src="../deps/images/goods/set of tea table and three chairs.jpg" class="card-img-top" alt="...">
+                <div class="card-body">
+                    <a href="">
+                        <p class="card-title">{{ product.name }}</p>
                     </a>
+                    <p class="card-text text-truncate">{{ product.description }}</p>
+                    <p class="product_id">{{ product.id }}</p>
+                    <div class="d-flex justify-content-between">
+<!--                        <p><s>100</s> P</p>-->
+                        <p><strong>{{ product.price }}</strong></p>
+<!--                        <span class="badge bg-warning text-dark">Скидка 10.00 %</span>-->
+<!--                         <p><strong>100.00 $</strong></p>-->
+                        <a href="{{ url_for('cart.cart') }}" class="btn add-to-cart">
+                            <img class="mx-1" src="{{ url_for('static', filename='img/cart.png') }}" alt="Catalog Icon" width="32" height="32">
+                        </a>
+                    </div>
                 </div>
             </div>
         </div>
-    </div>
+    {% endfor %}
+</div>
 
-        {% endfor %}
-    </div>
-    <nav aria-label="Page navigation example">
-        <ul class="pagination justify-content-center my-4">
-            <div class="custom-shadow d-flex">
-                <li class="page-item disabled">
-                    <a class="page-link">Пред</a>
-                </li>
-                <li class="page-item"><a class="page-link" href="#">1</a></li>
-                <li class="page-item"><a class="page-link" href="#">2</a></li>
-                <li class="page-item"><a class="page-link" href="#">3</a></li>
-                <li class="page-item">
-                    <a class="page-link" href="#">След</a>
-                </li>
-            </div>
-        </ul>
-    </nav>
-    {{ content }}
-    {% endblock %}
+{{ content }}
+{% endblock %}
