+++ app/user/models.py
@@ -2,7 +2,7 @@ from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 from datetime import datetime
 
-from app.models import db
+from app.db import db
 
 class User(db.Model, UserMixin):
     __tablename__ = 'users'
